Extrct part of the logic to separate files.
